requirements.txt
certifi==2021.5.30
click==8.0.3
dataclasses==0.8
Flask==2.0.2
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
kafka-python==2.0.2
MarkupSafe==2.0.1
numpy==1.19.5
opencv-python==4.5.4.60
PyMySQL==1.0.2
scipy==1.5.4
typing_extensions==4.0.0
Werkzeug==2.0.2
zipp==3.6.0
